LayoutTests/imported/w3c/web-platform-tests/css/css-values/attr-serialization-expected.txt


FAIL 'attr(data-foo type(<length>))' as a computed value should serialize as '10px'. assert_equals: 'attr(data-foo type(<length>))' and '10px' should serialize the same in computed values. expected "10px" but got "784px"
